requirements.txt
# Requirements for running an optimized retrieval‑augmented generation (RAG)
# system with MLX on Apple silicon.
#
# This list includes:
#  • **MLX** and **mlx‑lm** for running and loading quantized LLMs from the
#    Hugging Face *mlx‑community* repository.
#  • **LangChain**, its community extensions and integrations to build chains,
#    connect Hugging Face models, and split text into chunks.  We omit Chroma
#    here because we use FAISS as the vector store.
#  • **Sentence Transformers** for generating dense embeddings for the
#    document chunks.
#  • **PyMuPDF** (pymupdf) for loading and parsing PDF documents.  We still
#    include **pypdf** for basic PDF operations and compatibility.
#  • **FAISS** for similarity search in the RAG pipeline (replacing Chroma).
#  • **Additional utilities** for performance monitoring, caching, and error handling.
#
# Versions are pinned to specific releases as of October 3, 2025 to ensure
# compatibility with Python 3.13 and stability across installations.

# Core MLX dependencies
mlx==0.29.2
mlx-lm==0.28.2

# LangChain ecosystem
langchain==0.3.27
langchain-community==0.3.30
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11

# Embeddings and vector search
sentence-transformers==5.1.1
faiss-cpu==1.12.0

# Document processing
pymupdf==1.26.4
pypdf==6.1.1

# Web API dependencies
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Additional utilities for optimization
typing-extensions==4.12.2
pathlib2==2.3.7; python_version < "3.4"

# GUI dependencies
streamlit==1.39.0

# Testing dependencies
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-httpserver==1.0.11
responses==0.25.3
requests-mock==1.12.1
freezegun==1.5.5
httpx==0.28.1
